Renaming development_environment to environment
